backend/requirements.txt
# Core
Django==5.0.6
djangorestframework==3.15.2
django-environ==0.11.2

# Database
psycopg2-binary==2.9.9

# Cache
django-redis==5.4.0
redis==5.0.8

# Auth
djangorestframework-simplejwt==5.3.1

# Scraper
beautifulsoup4==4.12.3
requests==2.32.3
apscheduler==3.10.4

# Image handling
Pillow==10.4.0

# CORS
django-cors-headers==4.4.0

# Dev tools
ipython==8.26.0
